Add Linux support Tested on Ubuntu 22.04 with GCC 13.1.0 - Remove configuration subdirectories (Debug, Release...) - Window and clipboard implementation (xcb) - Specialize type_hash of ECS components (compiler independent value) - Use POSIX style path separator - Update RmlUi to master (vcpkg package did not work) - Remove local copy of glad (use github repo)

+++ modules/Audio/AudioDevice/include/audio/Utility.hpp
@@ -1,16 +1,17 @@
 #pragma once
 
 #include "glm/common.hpp" // clamp
+#include <cmath>
 
 namespace audio {
 
 [[nodiscard]] inline float dBToVolume(const int8_t dB) {
-  return std::powf(10.0f, 0.05f * dB);
+  return std::pow(10.0f, 0.05f * dB);
 }
 // @param volume [0..1]
 [[nodiscard]] inline float volumeTodB(float volume) {
   volume = glm::clamp(volume, 0.001f, 1.0f); // -60dB..0dB
-  return 20.0f * std::log10f(volume);
+  return 20.0f * std::log10(volume);
 }
 
 } // namespace audio
